requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
av==14.0.1
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
blessed==1.20.0
blinker==1.9.0
came-pytorch==0.1.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dashscope==1.22.1
datasets==3.2.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.2
diffusers==0.32.1
dill==0.3.8
DistVAE==0.0.0b5
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
flash_attn==2.7.4.post1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
func_timeout==4.3.5
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
gradio==3.48.0
gradio_client==0.6.1
grpcio==1.68.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.19.1
networkx==3.4.2
ninja==1.11.1.3
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime-gpu==1.19.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum-quanto==0.2.6
orjson==3.10.12
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
ray==2.42.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rpds-py==0.22.3
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
starlette==0.41.3
stringzilla==3.11.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.12.12
timm==1.0.12
tokenizers==0.21.0
tomesd==0.1.3
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torch-optimi==0.2.1
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.20.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.49.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.7
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
xformers==0.0.28.post3
xfuser==0.4.1
xxhash==3.5.0
yarl==1.18.3
yunchang==0.6.0
zipp==3.21.0
